--- packages/extension-browser/src/devtools/views/controls/settings.css
@@ -1,3 +1,0 @@
-.root {
-    margin-top: 1.5rem;
-}
